+++ tools/who_what_benchmark/tests/test_unit_speech_recognition.py
@@ -10,7 +10,7 @@ import pytest
 import torch
 
 from whowhatbench.whowhat_metrics import WordSimilarity
-from whowhatbench.wwb import to_mono_16k
+from whowhatbench.utils import to_mono_16k
 from whowhatbench.speech_recognition_evaluator import (
     FunASROptimumTranscriber,
     SpeechRecognitionEvaluator,
